rename all builders as "builder"
This is technically a breaking change, but cleans up the conventions and
provides improved organization.
